Types are short, not int.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@726 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/NXTRecord.java
@@ -76,7 +76,7 @@ rdataToString() {
 	if (next != null) {
 		sb.append(next);
 		int length = BitSetLength(bitmap);
-		for (int i = 0; i < length; i++)
+		for (short i = 0; i < length; i++)
 			if (bitmap.get(i)) {
 				sb.append(" ");
 				sb.append(Type.string(i));
